requirements.txt
# ── Core LLM / Agent framework ────────────────────────────────────────────────
langchain==0.2.0
langgraph==0.1.1
langsmith==0.1.77
langchain-community==0.2.0

# ── LLM serving clients ───────────────────────────────────────────────────────
httpx==0.27.0
openai==1.30.0              # vLLM uses OpenAI-compatible API

# ── Kubeflow Pipelines ────────────────────────────────────────────────────────
kfp==2.4.0
kfp-kubernetes==1.0.0
urllib3==1.26.20             # KFP compat (not 2.x)

# ── RAG / Vector ──────────────────────────────────────────────────────────────
qdrant-client==1.8.0
sentence-transformers==2.7.0
ragas==0.1.7
datasets==2.19.0

# ── Storage ───────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9
redis==5.0.4
sqlalchemy==2.0.29
boto3==1.34.69
s3fs==2024.3.1

# ── MLOps ─────────────────────────────────────────────────────────────────────
mlflow==2.12.2
dvc[s3]==3.49.0
evidently==0.4.22
great-expectations==0.18.12

# ── API / Serving ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-multipart==0.0.9

# ── Monitoring / Observability ────────────────────────────────────────────────
prometheus-client==0.20.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.1.0
pytest-asyncio==0.23.6
pytest-timeout==2.3.1
locust==2.28.0
httpx==0.27.0               # also used in tests

# ── Kubernetes client ─────────────────────────────────────────────────────────
kubernetes==26.1.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pyyaml==6.0.1
tenacity==8.3.0
requests==2.32.0
